Moved arrow key handlers to the body view
