+++ addon/views/llama-body.js
@@ -1,11 +1,14 @@
 import Em from 'ember';
 import ScrollXYMixin from 'llama-table/mixins/scroll-xy';
+import ArrowKeysMixin from 'llama-table/mixins/arrow-keys';
 import CopyController from 'llama-table/controllers/copy';
 var get = Em.get;
 var set = Em.set;
 
-var LlamaBody = Em.CollectionView.extend(ScrollXYMixin, {
+var LlamaBody = Em.CollectionView.extend(ScrollXYMixin, ArrowKeysMixin, {
 	classNames: 'llama-body',
+	attributeBindings: ['tabindex'],
+	tabindex: 0,
 	itemViewClass: Em.computed.alias('controller.BodyColumngroupView'),
 	columngroupViews: Em.computed.alias('childViews'),
 
@@ -30,6 +33,14 @@ var LlamaBody = Em.CollectionView.extend(ScrollXYMixin, {
 		if (isCopy && !isRange) {
 			controller.copy(e.target);
 		}
+		else {
+			this._super(e);
+		}
+	},
+
+	focusIn: function (e) {
+		// TODO: prevents shift-tab escaping
+		this.get('controller').send('focusCurrentCell');
 	},
 
 	actions: {
@@ -38,6 +49,18 @@ var LlamaBody = Em.CollectionView.extend(ScrollXYMixin, {
 		},
 		scrollY: function (pos) {
 			this.get('controller').send('scrollY', pos);
+		},
+		keyLeft: function () {
+			this.get('controller').send('focusLeft');
+		},
+		keyUp: function () {
+			this.get('controller').send('focusUp');
+		},
+		keyRight: function () {
+			this.get('controller').send('focusRight');
+		},
+		keyDown: function () {
+			this.get('controller').send('focusDown');
 		}
 	}
 });
